Refine dependency management glue
Separates data access interface (now a component) from the main module.
This approach is closer to what the code would look like with Dagger.

Also splitting data access and use cases into separate dependencies is
more in-line with the testing strategy since it allows the application
to specify which data source to use, but not specify mock use cases.