Don't make distinction between disk and network IO
The thread pool allows for a fair number of active threads so a few in-progress
network requests won't block disk IO work. If the app eventually uses a lot of
concurrent network requests, it may sense to make the pool unbounded so that
the slow network requests don't block disk IO access.